refactor: creating helper file with helper functions used in multiple places
